demo meet fixes

+++ src/AuctionScreen/Screens/profile/AuctionProfile.js
@@ -19,7 +19,7 @@ import { Media } from '../../../Global/Media';
 import { Poppins } from '../../../Global/FontFamily';
 import { Iconviewcomponent } from '../../../Components/Icontag';
 
-import { useNavigation, useRoute } from '@react-navigation/native';
+import { useFocusEffect, useNavigation, useRoute } from '@react-navigation/native';
 import { useSelector, useDispatch } from 'react-redux';
 import { setActionUserData, setAsync, setLoginType, setUserData } from '../../../Redux';
 import AsyncStorage from '@react-native-async-storage/async-storage';
@@ -31,6 +31,7 @@ import AuctionBottomLogin from '../../Auctioncomponents/AuctionBottomLogin';
 import { Linking } from 'react-native';
 import moment from 'moment';
 import common_fn from '../../../Config/common_fn';
+import { baseUrl } from '../../../Config/base_url';
 var { width, height } = Dimensions.get('screen');
 
 const AuctionProfile = ({ navigation }) => {
@@ -83,7 +84,15 @@ const AuctionProfile = ({ navigation }) => {
       return () => clearInterval(interval);
     }
   }, [data?.id]);
-
+  useFocusEffect(
+    React.useCallback(() => {
+      setLoading(true);
+      profileShowData();
+      const interval = setTimeout(() => {
+        setLoading(false);
+      }, 1000);
+      return () => clearInterval(interval);
+    }, []))
 
   useEffect(() => {
     try {
@@ -101,7 +110,7 @@ const AuctionProfile = ({ navigation }) => {
         redirect: "follow"
       };
       // console.log("Profile user id =========== :", data?.id);
-      fetch(`https://api.albionbankauctions.com/api/user/get_user?user_id=${data?.id}&status=activated`, requestOptions)
+      fetch(`${baseUrl}api/user/get_user?user_id=${data?.id}&status=activated`, requestOptions)
         .then((response) => response.json())
         .then((result) => {
           if (result?.status == true) {
@@ -109,7 +118,6 @@ const AuctionProfile = ({ navigation }) => {
           } else {
             common_fn.showToast(result?.message);
           }
-          console.log("Profile Data =========== :", result.data);
         })
         .catch((error) => console.error("catch in profileShowData_Api :", error));
     } catch (error) {
@@ -130,14 +138,13 @@ const AuctionProfile = ({ navigation }) => {
       };
 
       // fetch(`http://192.168.29.204:5000/api/plan/user?user_id=${id}`, requestOptions)
-      fetch(`http://13.127.95.5:5000/api/plan/user?user_id=${data?.id}&status=activated`, requestOptions)
+      fetch(`${baseUrl}api/plan/user?user_id=${data?.id}&status=activated`, requestOptions)
         .then((response) => response.json())
         .then((result) => {
           if (result?.status == true) {
             // console.log("profile data -----------------113232123", result?.data[0]?.Invoice)
             setExpireDate(moment(result?.data[0]?.expires_at).format('DD-MM-YYYY'));
-            console.log("http://13.127.95.5:5000/api/plan/2266",result?.data[0]);
-            
+
             setPlanStatus(result?.data[0]?.plan_id)
             setInvoiceShown(result?.data[0]?.Invoice)
             // console.log("PLAN ======= :", result?.data[0])
@@ -150,10 +157,6 @@ const AuctionProfile = ({ navigation }) => {
       console.log("catch in plan_CheckData_Home : ", error);
     }
   }
-
-  console.log("data?.id ============== :", data?.id);
-
-
 
   return (
     <View style={{ flex: 1, backgroundColor: Color.white }}>
@@ -224,7 +227,7 @@ const AuctionProfile = ({ navigation }) => {
                 style={{
                   justifyContent: 'center',
                   alignItems: 'center',
-                  marginVertical: 10,
+                  marginVertical: 2,
                 }}>
                 {data?.id != undefined ?
                   <Text
@@ -241,10 +244,10 @@ const AuctionProfile = ({ navigation }) => {
                 {data?.id != undefined ?
                   <Text
                     style={{
-                      fontSize: 14,
+                      fontSize: 13,
                       color: '#444',
-                      fontFamily: Poppins.SemiBold,
-                      marginVertical: 5,
+                      fontFamily: Poppins.Medium,
+                      // marginVertical: 5,
                     }}>
                     {userData?.phone_number}
                   </Text>
@@ -255,30 +258,39 @@ const AuctionProfile = ({ navigation }) => {
                     style={{
                       fontSize: 13,
                       color: Color.black,
-                      fontFamily: Poppins.Bold,
+                      fontFamily: Poppins.SemiBold,
+                      marginBottom: 5
                     }}>
                     {userData?.address}
                   </Text> : null}
-
-                {data?.id != undefined ?
+                {data?.id != undefined && planStatus ?
                   <Text
                     style={{
                       fontSize: 12,
                       color: Color.white,
-                      fontFamily: Poppins.Bold, backgroundColor: Color.primary, borderRadius: 30, padding: 7, paddingHorizontal: 30, marginVertical: 10
+                      fontFamily: Poppins.Medium, backgroundColor: Color.primary, borderRadius: 30, padding: 7, paddingHorizontal: 30, marginBottom: 10
                     }}>
                     {planStatus == 1 ? "Free Plan" : planStatus == 2 ? "3 Months Plan" : planStatus == 3 ? "6 Months Plan" : planStatus == 4 ? "12 Months Plan" : null}
-                  </Text> : null}
+                  </Text> : <TouchableOpacity style={{
+                    fontSize: 12,
+                    color: Color.white,
+                    marginBottom: 10,
+                    fontFamily: Poppins.Bold, backgroundColor: Color.primary, borderRadius: 30, padding: 7, paddingHorizontal: 30
+                  }} onPress={() => navigation.navigate('AuctionPrime')}>
+                    <Text style={{ color: Color.white, fontFamily: Poppins.Medium }}>
+                      Upgrade plan
+                    </Text>
+                  </TouchableOpacity>}
 
 
-                {data?.id != undefined && expireDate != "null" ?
+                {data?.id != undefined && expireDate != "null" && planStatus ?
                   <Text
                     style={{
                       fontSize: 12,
                       color: Color.black,
                       fontFamily: Poppins.Bold,
                     }}>
-                    Expired Date : {expireDate}
+                    Expires At : {expireDate}
                   </Text>
                   : null}
               </View>
@@ -311,11 +323,11 @@ const AuctionProfile = ({ navigation }) => {
                       marginHorizontal: 10,
                     }}>
                     <View style={{ flexDirection: 'row', alignItems: 'center' }}>
-                      <Text style={{ fontSize: 14, color: '#333', fontWeight: 'bold' }}>
+                      <Text style={{ fontSize: 14, color: '#333', fontFamily: Poppins?.Medium }}>
                         Profile
                       </Text>
                     </View>
-                    <Text style={{ fontSize: 12, color: '#666' }}>
+                    <Text style={{ fontSize: 12, color: '#666', fontFamily: Poppins?.Medium }}>
                       Your profile details will be edited here
                     </Text>
                   </View>
@@ -347,11 +359,11 @@ const AuctionProfile = ({ navigation }) => {
                       marginHorizontal: 10,
                     }}>
                     <View style={{ flexDirection: 'row', alignItems: 'center' }}>
-                      <Text style={{ fontSize: 14, color: '#333', fontWeight: 'bold' }}>
+                      <Text style={{ fontSize: 14, color: '#333', fontFamily: Poppins?.Medium }}>
                         Interested properties
                       </Text>
                     </View>
-                    <Text style={{ fontSize: 12, color: '#666' }}>
+                    <Text style={{ fontSize: 12, color: '#666', fontFamily: Poppins?.Medium }}>
                       Your Interested property details
                     </Text>
                   </View>
@@ -384,11 +396,11 @@ const AuctionProfile = ({ navigation }) => {
                       marginHorizontal: 10,
                     }}>
                     <View style={{ flexDirection: 'row', alignItems: 'center' }}>
-                      <Text style={{ fontSize: 14, color: '#333', fontWeight: 'bold' }}>
+                      <Text style={{ fontSize: 14, color: '#333', fontFamily: Poppins?.Medium }}>
                         Subscription Details
                       </Text>
                     </View>
-                    <Text style={{ fontSize: 12, color: '#666' }}>
+                    <Text style={{ fontSize: 12, color: '#666', fontFamily: Poppins?.Medium }}>
                       Your subscription plan details
                     </Text>
                   </View>
@@ -427,11 +439,11 @@ const AuctionProfile = ({ navigation }) => {
                           marginHorizontal: 10,
                         }}>
                         <View style={{ flexDirection: 'row', alignItems: 'center' }}>
-                          <Text style={{ fontSize: 14, color: '#333', fontWeight: 'bold' }}>
+                          <Text style={{ fontSize: 14, color: '#333', fontFamily: Poppins?.Medium }}>
                             Invoice List
                           </Text>
                         </View>
-                        <Text style={{ fontSize: 12, color: '#666' }}>
+                        <Text style={{ fontSize: 12, color: '#666', fontFamily: Poppins?.Medium }}>
                           Your invoice list details
                         </Text>
                       </View>
@@ -446,82 +458,46 @@ const AuctionProfile = ({ navigation }) => {
                   </View>
                   :
                   null}
-                {data?.id == undefined ? (
-                  <View />
-                ) : (
                   <TouchableOpacity
-                    onPress={() => {
-                      Alert.alert(
-                        '',
-                        'Do You like to remove your account',
-                        [
-                          {
-                            text: 'No',
-                            onPress: async () => { },
-                          },
-                          {
-                            text: 'Yes',
-                            onPress: async () => {
-                              Linking.openURL(`https://albionbankauctions.com/web/user/delete/${data?.id}`,);
-                              AsyncStorage.clear();
-                              navigation.replace('OnboardingScreen2');
-                              dispatch(setActionUserData({}));
-                              dispatch(setLoginType(''));
-                              // const usersData = await fetchData.Auction_deleteData(
-                              //   data,
-                              // );
-                              // if (usersData?.message == 'Success') {
-                              //   AsyncStorage.clear();
-                              //   navigation.replace('ActionLogin');
-                              //   dispatch(setActionUserData({}));
-                              //   dispatch(setLoginType(''));
-                              // }
-
-                            },
-                          },
-                        ],
-                        { cancelable: false },
-                      );
-                    }}
+                  style={{
+                    flexDirection: 'row',
+                    alignItems: 'center',
+                    justifyContent: 'center',
+                    marginVertical: 10,
+                    marginHorizontal: 10,
+                  }}
+                  onPress={() => {
+                    navigation.navigate('DeleteProfile',{
+                      data: data
+                    })
+                  }}
+                >
+                  <Iconviewcomponent
+                    Icontag={'Ionicons'}
+                    iconname={'options-outline'}
+                    icon_size={24}
+                    icon_color={Color.primary}
+                  />
+                  <View
                     style={{
-                      flexDirection: 'row',
-                      alignItems: 'center',
-                      justifyContent: 'center',
-                      marginVertical: 10,
+                      flex: 1,
+                      justifyContent: 'flex-start',
+                      alignItems: 'flex-start',
                       marginHorizontal: 10,
                     }}>
-                    {/* <Icon name="heart-outline" size={25} color={Color.primary} /> */}
-                    <Iconviewcomponent
-                      Icontag={'AntDesign'}
-                      iconname={'deleteuser'}
-                      icon_size={22}
-                      icon_color={Color.primary}
-                    />
-                    <View
-                      style={{
-                        flex: 1,
-                        justifyContent: 'flex-start',
-                        alignItems: 'flex-start',
-                        marginHorizontal: 10,
-                      }}>
-                      <Text style={{ fontSize: 14, color: '#333', fontWeight: 'bold' }}>
-                        Delete User
-                      </Text>
-                      <Text style={{ fontSize: 12, color: '#666' }}>
-                        Removing Your Account
-                      </Text>
-                    </View>
-                    <Iconviewcomponent
-                      Icontag={'Ionicons'}
-                      iconname={'chevron-forward'}
-                      icon_size={18}
-                      icon_color={'#666'}
-                    />
-                  </TouchableOpacity>
-                )}
-
+                    <Text
+                      style={{ fontSize: 14, color: '#333', fontFamily: Poppins?.Medium }}>
+                      Others
+                    </Text>
+                  </View>
+                  <Iconviewcomponent
+                    Icontag={'Ionicons'}
+                    iconname={'chevron-forward'}
+                    icon_size={18}
+                    icon_color={'#666'}
+                  />
+                </TouchableOpacity>
                 <Divider style={{ height: 1, marginVertical: 10 }} />
-
                 <TouchableOpacity
                   style={{
                     flexDirection: 'row',
@@ -530,26 +506,19 @@ const AuctionProfile = ({ navigation }) => {
                     marginVertical: 10,
                     marginHorizontal: 10,
                   }}
-                  // onPress={async () => {
-                  //   try {
-                  //     await AsyncStorage.removeItem('user_data');  // ✅ Remove stored user data
-
-                  //     dispatch(setUserData(null)); // ✅ Reset Redux state
-                  //     dispatch(setActionUserData(null));
-                  //     dispatch(setLoginType(''));
-
-                  //     navigation.replace('OnboardingScreen2');  // ✅ Navigate to onboarding
-
-                  //   } catch (error) {
-                  //     console.error("Logout Error:", error);
-                  //   }
-                  // }}
-
                   onPress={async () => {
-                    await AsyncStorage.clear();
-                    navigation.navigate('OnboardingScreen2');
-                    // dispatch(setActionUserData({}));
-                    // dispatch(setLoginType(''));
+                    console.log("userid", data?.id);
+                    const logout = await fetchData?.Auction_Logout_Api(data?.id);
+                    console.log("logout", logout);
+                    if (logout?.message == "Logout Successfully") {
+                      await AsyncStorage.clear();
+                      common_fn.showToast('Logout Successfully');
+                      navigation.navigate('OnboardingScreen2');
+                      navigation.reset({
+                        index: 0,
+                        routes: [{ name: 'OnboardingScreen2' }],
+                      })
+                    }
                   }}
                 >
                   <Iconviewcomponent
@@ -566,7 +535,7 @@ const AuctionProfile = ({ navigation }) => {
                       marginHorizontal: 10,
                     }}>
                     <Text
-                      style={{ fontSize: 14, color: '#333', fontWeight: 'bold' }}>
+                      style={{ fontSize: 14, color: '#333', fontFamily: Poppins?.Medium }}>
                       Logout
                     </Text>
                   </View>
@@ -577,7 +546,6 @@ const AuctionProfile = ({ navigation }) => {
                     icon_color={'#666'}
                   />
                 </TouchableOpacity>
-                <Divider style={{ height: 1, marginVertical: 10 }} />
               </View>
               : null}
